requirements.txt
streamlit 
langchain 
openai
langchain_openai
python-dotenv
datasets
langchain_community
chromadb
protobuf==3.20.3
pysqlite3-binary